requirements.txt
flask==1.0.2
slackclient==2.0.0
boto3
bs4